feat: Remove ENABLE_FORUM_V2 waffle flag and update tests (#652)
