+++ lms/djangoapps/discussion/tests/test_tasks_v2.py
@@ -123,7 +123,7 @@ class TaskTestCase(
         """mock threads and comments"""
         if subscribed_thread_ids:
             self.set_mock_side_effect(
-                "get_user_subscriptions",
+                "get_user_threads",
                 make_subscribed_threads_callback(subscribed_thread_ids, per_page),
             )
         if thread_data:
