+++ lms/djangoapps/discussion/tests/test_views.py
@@ -1,3 +1,4 @@
+# pylint: disable=unused-import
 """
 Tests the forum notification views.
 """
@@ -91,12 +92,6 @@ class ViewsExceptionTestCase(UrlResetMixin, ModuleStoreTestCase):  # lint-amnest
         config = ForumsConfig.current()
         config.enabled = True
         config.save()
-        patcher = mock.patch(
-            'openedx.core.djangoapps.discussions.config.waffle.ENABLE_FORUM_V2.is_enabled',
-            return_value=False
-        )
-        patcher.start()
-        self.addCleanup(patcher.stop)
         patcher = mock.patch(
             "openedx.core.djangoapps.django_comment_common.comment_client.thread.forum_api.get_course_id_by_thread"
         )
@@ -337,12 +332,6 @@ class CommentsServiceRequestHeadersTestCase(ForumsEnableMixin, UrlResetMixin, Mo
     def setUp(self):
         super().setUp()
         patcher = mock.patch(
-            'openedx.core.djangoapps.discussions.config.waffle.ENABLE_FORUM_V2.is_enabled',
-            return_value=False
-        )
-        patcher.start()
-        self.addCleanup(patcher.stop)
-        patcher = mock.patch(
             "openedx.core.djangoapps.django_comment_common.comment_client.models.forum_api.get_course_id_by_comment"
         )
         self.mock_get_course_id_by_comment = patcher.start()
@@ -407,36 +396,6 @@ class CommentsServiceRequestHeadersTestCase(ForumsEnableMixin, UrlResetMixin, Mo
             ),
         )
         self.assert_all_calls_have_header(mock_request, "X-Edx-Api-Key", "test_api_key")
-
-
-class UserProfileUnicodeTestCase(ForumsEnableMixin, SharedModuleStoreTestCase, UnicodeTestMixin):  # lint-amnesty, pylint: disable=missing-class-docstring
-
-    @classmethod
-    def setUpClass(cls):
-        # pylint: disable=super-method-not-called
-        with super().setUpClassAndTestData():
-            cls.course = CourseFactory.create()
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-
-        cls.student = UserFactory.create()
-        CourseEnrollmentFactory(user=cls.student, course_id=cls.course.id)
-
-    @patch('openedx.core.djangoapps.django_comment_common.comment_client.utils.requests.request', autospec=True)
-    def _test_unicode_data(self, text, mock_request):  # lint-amnesty, pylint: disable=missing-function-docstring
-        mock_request.side_effect = make_mock_request_impl(course=self.course, text=text)
-        request = RequestFactory().get("dummy_url")
-        request.user = self.student
-        # so (request.headers.get('x-requested-with') == 'XMLHttpRequest') == True
-        request.META["HTTP_X_REQUESTED_WITH"] = "XMLHttpRequest"
-
-        response = views.user_profile(request, str(self.course.id), str(self.student.id))
-        assert response.status_code == 200
-        response_data = json.loads(response.content.decode('utf-8'))
-        assert response_data['discussion_data'][0]['title'] == text
-        assert response_data['discussion_data'][0]['body'] == text
 
 
 class EnrollmentTestCase(ForumsEnableMixin, ModuleStoreTestCase):
@@ -746,76 +705,3 @@ class DefaultTopicIdGetterTestCase(ModuleStoreTestCase):
         expected_id = 'another_discussion_id'
         result = _get_discussion_default_topic_id(course)
         assert expected_id == result
-
-
-@ddt.ddt
-@patch(
-    'openedx.core.djangoapps.django_comment_common.comment_client.utils.perform_request',
-    Mock(
-        return_value={
-            "id": "test_thread",
-            "title": "Title",
-            "body": "<p></p>",
-            "default_sort_key": "date",
-            "upvoted_ids": [],
-            "downvoted_ids": [],
-            "subscribed_thread_ids": [],
-        }
-    )
-)
-class ForumMFETestCase(ForumsEnableMixin, SharedModuleStoreTestCase):
-    """
-    Tests that the MFE upgrade banner and MFE is shown in the correct situation with the correct UI
-    """
-
-    def setUp(self):
-        super().setUp()
-        self.course = CourseFactory.create()
-        self.user = UserFactory.create()
-        self.staff_user = AdminFactory.create()
-        CourseEnrollmentFactory.create(user=self.user, course_id=self.course.id)
-
-    @override_settings(DISCUSSIONS_MICROFRONTEND_URL="http://test.url")
-    def test_redirect_from_legacy_base_url_to_new_experience(self):
-        """
-        Verify that the legacy url is redirected to MFE homepage when
-        ENABLE_DISCUSSIONS_MFE flag is enabled.
-        """
-
-        with override_waffle_flag(ENABLE_DISCUSSIONS_MFE, True):
-            self.client.login(username=self.user.username, password=self.TEST_PASSWORD)
-            url = reverse("forum_form_discussion", args=[self.course.id])
-            response = self.client.get(url)
-            assert response.status_code == 302
-            expected_url = f"{settings.DISCUSSIONS_MICROFRONTEND_URL}/{str(self.course.id)}"
-            assert response.url == expected_url
-
-    @override_settings(DISCUSSIONS_MICROFRONTEND_URL="http://test.url")
-    def test_redirect_from_legacy_profile_url_to_new_experience(self):
-        """
-        Verify that the requested user profile is redirected to MFE learners tab when
-        ENABLE_DISCUSSIONS_MFE flag is enabled
-        """
-
-        with override_waffle_flag(ENABLE_DISCUSSIONS_MFE, True):
-            self.client.login(username=self.user.username, password=self.TEST_PASSWORD)
-            url = reverse("user_profile", args=[self.course.id, self.user.id])
-            response = self.client.get(url)
-            assert response.status_code == 302
-            expected_url = f"{settings.DISCUSSIONS_MICROFRONTEND_URL}/{str(self.course.id)}/learners"
-            assert response.url == expected_url
-
-    @override_settings(DISCUSSIONS_MICROFRONTEND_URL="http://test.url")
-    def test_redirect_from_legacy_single_thread_to_new_experience(self):
-        """
-        Verify that a legacy single url is redirected to corresponding MFE thread url when the ENABLE_DISCUSSIONS_MFE
-        flag is enabled
-        """
-
-        with override_waffle_flag(ENABLE_DISCUSSIONS_MFE, True):
-            self.client.login(username=self.user.username, password=self.TEST_PASSWORD)
-            url = reverse("single_thread", args=[self.course.id, "test_discussion", "test_thread"])
-            response = self.client.get(url)
-            assert response.status_code == 302
-            expected_url = f"{settings.DISCUSSIONS_MICROFRONTEND_URL}/{str(self.course.id)}/posts/test_thread"
-            assert response.url == expected_url
