+++ lms/djangoapps/discussion/tests/test_views_v2.py
@@ -286,7 +286,6 @@ class ForumViewsUtilsMixin(MockForumApiMixin):
             "search_threads",
             "get_user_active_threads",
             "get_user_threads",
-            "get_user_subscriptions",
         ]:
             self.set_mock_side_effect(
                 func_name,
@@ -847,7 +846,7 @@ class SingleThreadContentGroupTestCase(ForumsEnableMixin, UrlResetMixin, Content
 
 
 class FollowedThreadsDiscussionGroupIdTestCase(CohortedTestCase, CohortedTopicGroupIdTestMixinV2, ForumViewsUtilsMixin):  # lint-amnesty, pylint: disable=missing-class-docstring
-    function_name = "get_user_subscriptions"
+    function_name = "get_user_threads"
 
     @classmethod
     def setUpClass(cls):
@@ -2022,3 +2021,115 @@ class FollowedThreadsUnicodeTestCase(
         response_data = json.loads(response.content.decode('utf-8'))
         assert response_data['discussion_data'][0]['title'] == text
         assert response_data['discussion_data'][0]['body'] == text
+
+
+class UserProfileUnicodeTestCase(ForumsEnableMixin, SharedModuleStoreTestCase, UnicodeTestMixin, ForumViewsUtilsMixin):  # lint-amnesty, pylint: disable=missing-class-docstring
+
+    @classmethod
+    def setUpClass(cls):
+        # pylint: disable=super-method-not-called
+        with super().setUpClassAndTestData():
+            cls.course = CourseFactory.create()
+        super().setUpClassAndForumMock()
+
+    @classmethod
+    def tearDownClass(cls):
+        super().tearDownClass()
+        super().disposeForumMocks()
+
+    @classmethod
+    def setUpTestData(cls):
+        super().setUpTestData()
+
+        cls.student = UserFactory.create()
+        CourseEnrollmentFactory(user=cls.student, course_id=cls.course.id)
+
+    def _test_unicode_data(self, text):  # lint-amnesty, pylint: disable=missing-function-docstring
+        self._configure_mock_responses(course=self.course, text=text)
+        request = RequestFactory().get("dummy_url")
+        request.user = self.student
+        # so (request.headers.get('x-requested-with') == 'XMLHttpRequest') == True
+        request.META["HTTP_X_REQUESTED_WITH"] = "XMLHttpRequest"
+
+        response = views.user_profile(request, str(self.course.id), str(self.student.id))
+        assert response.status_code == 200
+        response_data = json.loads(response.content.decode('utf-8'))
+        assert response_data['discussion_data'][0]['title'] == text
+        assert response_data['discussion_data'][0]['body'] == text
+
+
+class ForumMFETestCase(ForumsEnableMixin, SharedModuleStoreTestCase, ModuleStoreTestCase, MockForumApiMixin):  # lint-amnesty, pylint: disable=missing-class-docstring
+    """
+    Tests that the MFE upgrade banner and MFE is shown in the correct situation with the correct UI
+    """
+
+    def setUp(self):
+        super().setUp()
+        self.course = CourseFactory.create()
+        self.user = UserFactory.create()
+        self.staff_user = AdminFactory.create()
+        CourseEnrollmentFactory.create(user=self.user, course_id=self.course.id)
+        self.set_mock_return_value("get_user", {
+            "id": "test_thread",
+            "title": "Title",
+            "body": "<p></p>",
+            "default_sort_key": "date",
+            "upvoted_ids": [],
+            "downvoted_ids": [],
+            "subscribed_thread_ids": [],
+        })
+        self.set_mock_return_value("get_user_active_threads", {})
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        super().setUpClassAndForumMock()
+
+    @classmethod
+    def tearDownClass(cls):
+        super().tearDownClass()
+        super().disposeForumMocks()
+
+    @override_settings(DISCUSSIONS_MICROFRONTEND_URL="http://test.url")
+    def test_redirect_from_legacy_base_url_to_new_experience(self):
+        """
+        Verify that the legacy url is redirected to MFE homepage when
+        ENABLE_DISCUSSIONS_MFE flag is enabled.
+        """
+
+        with override_waffle_flag(ENABLE_DISCUSSIONS_MFE, True):
+            self.client.login(username=self.user.username, password=self.TEST_PASSWORD)
+            url = reverse("forum_form_discussion", args=[self.course.id])
+            response = self.client.get(url)
+            assert response.status_code == 302
+            expected_url = f"{settings.DISCUSSIONS_MICROFRONTEND_URL}/{str(self.course.id)}"
+            assert response.url == expected_url
+
+    @override_settings(DISCUSSIONS_MICROFRONTEND_URL="http://test.url")
+    def test_redirect_from_legacy_profile_url_to_new_experience(self):
+        """
+        Verify that the requested user profile is redirected to MFE learners tab when
+        ENABLE_DISCUSSIONS_MFE flag is enabled
+        """
+        with override_waffle_flag(ENABLE_DISCUSSIONS_MFE, True):
+            self.client.login(username=self.user.username, password=self.TEST_PASSWORD)
+            url = reverse("user_profile", args=[self.course.id, self.user.id])
+            response = self.client.get(url)
+            assert response.status_code == 302
+            expected_url = f"{settings.DISCUSSIONS_MICROFRONTEND_URL}/{str(self.course.id)}/learners"
+            assert response.url == expected_url
+
+    @override_settings(DISCUSSIONS_MICROFRONTEND_URL="http://test.url")
+    def test_redirect_from_legacy_single_thread_to_new_experience(self):
+        """
+        Verify that a legacy single url is redirected to corresponding MFE thread url when the ENABLE_DISCUSSIONS_MFE
+        flag is enabled
+        """
+
+        with override_waffle_flag(ENABLE_DISCUSSIONS_MFE, True):
+            self.client.login(username=self.user.username, password=self.TEST_PASSWORD)
+            url = reverse("single_thread", args=[self.course.id, "test_discussion", "test_thread"])
+            response = self.client.get(url)
+            assert response.status_code == 302
+            expected_url = f"{settings.DISCUSSIONS_MICROFRONTEND_URL}/{str(self.course.id)}/posts/test_thread"
+            assert response.url == expected_url
