+++ openedx/core/djangoapps/discussions/config/waffle.py
@@ -2,8 +2,6 @@
 This module contains various configuration settings via
 waffle switches for the discussions app.
 """
-from django.conf import settings
-
 from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
 
 WAFFLE_FLAG_NAMESPACE = "discussions"
@@ -45,31 +43,3 @@ ENABLE_PAGES_AND_RESOURCES_MICROFRONTEND = CourseWaffleFlag(
 ENABLE_NEW_STRUCTURE_DISCUSSIONS = CourseWaffleFlag(
     f"{WAFFLE_FLAG_NAMESPACE}.enable_new_structure_discussions", __name__
 )
-
-# .. toggle_name: discussions.enable_forum_v2
-# .. toggle_implementation: CourseWaffleFlag
-# .. toggle_default: False
-# .. toggle_description: Waffle flag to use the forum v2 instead of v1(cs_comment_service)
-# .. toggle_use_cases: temporary, open_edx
-# .. toggle_creation_date: 2024-9-26
-# .. toggle_target_removal_date: 2025-12-05
-ENABLE_FORUM_V2 = CourseWaffleFlag(f"{WAFFLE_FLAG_NAMESPACE}.enable_forum_v2", __name__)
-
-
-def is_forum_v2_enabled(course_key):
-    """
-    Returns whether forum V2 is enabled on the course. This is a 2-step check:
-
-    1. Check value of settings.DISABLE_FORUM_V2: if it exists and is true, this setting overrides any course flag.
-    2. Else, check the value of the corresponding course waffle flag.
-    """
-    if is_forum_v2_disabled_globally():
-        return False
-    return ENABLE_FORUM_V2.is_enabled(course_key)
-
-
-def is_forum_v2_disabled_globally() -> bool:
-    """
-    Return True if DISABLE_FORUM_V2 is defined and true-ish.
-    """
-    return getattr(settings, "DISABLE_FORUM_V2", False)
