+++ openedx/core/djangoapps/django_comment_common/comment_client/course.py
@@ -8,9 +8,6 @@ from edx_django_utils.monitoring import function_trace
 from opaque_keys.edx.keys import CourseKey
 
 from forum import api as forum_api
-from openedx.core.djangoapps.django_comment_common.comment_client import settings
-from openedx.core.djangoapps.django_comment_common.comment_client.utils import perform_request
-from openedx.core.djangoapps.discussions.config.waffle import is_forum_v2_enabled
 
 
 def get_course_commentable_counts(course_key: CourseKey) -> Dict[str, Dict[str, int]]:
@@ -31,19 +28,7 @@ def get_course_commentable_counts(course_key: CourseKey) -> Dict[str, Dict[str, 
             }
 
     """
-    if is_forum_v2_enabled(course_key):
-        commentable_stats = forum_api.get_commentables_stats(str(course_key))
-    else:
-        url = f"{settings.PREFIX}/commentables/{course_key}/counts"
-        commentable_stats = perform_request(
-            'get',
-            url,
-            metric_tags=[
-                f"course_key:{course_key}",
-                "function:get_course_commentable_counts",
-            ],
-            metric_action='commentable_stats.retrieve',
-        )
+    commentable_stats = forum_api.get_commentables_stats(str(course_key))
     return commentable_stats
 
 
@@ -81,20 +66,7 @@ def get_course_user_stats(course_key: CourseKey, params: Optional[Dict] = None) 
     """
     if params is None:
         params = {}
-    if is_forum_v2_enabled(course_key):
-        course_stats = forum_api.get_user_course_stats(str(course_key), **params)
-    else:
-        url = f"{settings.PREFIX}/users/{course_key}/stats"
-        course_stats = perform_request(
-            'get',
-            url,
-            params,
-            metric_action='user.course_stats',
-            metric_tags=[
-                f"course_key:{course_key}",
-                "function:get_course_user_stats",
-            ],
-        )
+    course_stats = forum_api.get_user_course_stats(str(course_key), **params)
     return course_stats
 
 
@@ -109,17 +81,5 @@ def update_course_users_stats(course_key: CourseKey) -> Dict:
     Returns:
         dict: data returned by API. Contains count of users updated.
     """
-    if is_forum_v2_enabled(course_key):
-        course_stats = forum_api.update_users_in_course(str(course_key))
-    else:
-        url = f"{settings.PREFIX}/users/{course_key}/update_stats"
-        course_stats = perform_request(
-            'post',
-            url,
-            metric_action='user.update_course_stats',
-            metric_tags=[
-                f"course_key:{course_key}",
-                "function:update_course_users_stats",
-            ],
-        )
+    course_stats = forum_api.update_users_in_course(str(course_key))
     return course_stats
