+++ openedx/core/djangoapps/django_comment_common/comment_client/models.py
@@ -2,11 +2,9 @@
 
 
 import logging
-import typing as t
 
 from .utils import CommentClientRequestError, extract, perform_request, get_course_key
 from forum import api as forum_api
-from openedx.core.djangoapps.discussions.config.waffle import is_forum_v2_enabled, is_forum_v2_disabled_globally
 
 log = logging.getLogger(__name__)
 
@@ -74,10 +72,11 @@ class Model:
     def _retrieve(self, *args, **kwargs):
         course_id = self.attributes.get("course_id") or kwargs.get("course_key")
         if not course_id:
-            _, course_id = is_forum_v2_enabled_for_comment(self.id)
+            course_id = forum_api.get_course_id_by_comment(self.id)
+        response = None
         if self.type == "comment":
             response = forum_api.get_parent_comment(comment_id=self.attributes["id"], course_id=course_id)
-        else:
+        if response is None:
             raise CommentClientRequestError("Forum v2 API call is missing")
         self._update_from_response(response)
 
@@ -206,18 +205,15 @@ class Model:
             request_params.update(params)
         course_id = self.attributes.get("course_id") or request_params.get("course_id")
         course_key = get_course_key(course_id)
-        if is_forum_v2_enabled(course_key):
-            response = None
-            if self.type == "comment":
-                response = self.handle_update_comment(request_params, str(course_key))
-            elif self.type == "thread":
-                response = self.handle_update_thread(request_params, str(course_key))
-            elif self.type == "user":
-                response = self.handle_update_user(request_params, str(course_key))
-            if response is None:
-                raise CommentClientRequestError("Forum v2 API call is missing")
-        else:
-            response = self.perform_http_put_request(request_params)
+        response = None
+        if self.type == "comment":
+            response = self.handle_update_comment(request_params, str(course_key))
+        elif self.type == "thread":
+            response = self.handle_update_thread(request_params, str(course_key))
+        elif self.type == "user":
+            response = self.handle_update_user(request_params, str(course_key))
+        if response is None:
+            raise CommentClientRequestError("Forum v2 API call is missing")
         return response
 
     def handle_update_user(self, request_params, course_id):
@@ -274,28 +270,6 @@ class Model:
         response = forum_api.update_thread(**request_data)
         return response
 
-    def perform_http_put_request(self, request_params):
-        url = self.url(action="put", params=self.attributes)
-        response = perform_request(
-            "put",
-            url,
-            request_params,
-            metric_tags=self._metric_tags,
-            metric_action="model.update",
-        )
-        return response
-
-    def perform_http_post_request(self):
-        url = self.url(action="post", params=self.attributes)
-        response = perform_request(
-            "post",
-            url,
-            self.initializable_attributes(),
-            metric_tags=self._metric_tags,
-            metric_action="model.insert",
-        )
-        return response
-
     def handle_create(self, params=None):
         course_id = self.attributes.get("course_id") or params.get("course_id")
         course_key = str(get_course_key(course_id))
@@ -348,22 +322,3 @@ class Model:
 
         response = forum_api.create_thread(**params)
         return response
-
-
-def is_forum_v2_enabled_for_comment(comment_id: str) -> tuple[bool, t.Optional[str]]:
-    """
-    Figure out whether we use forum v2 for a given comment.
-
-    See is_forum_v2_enabled_for_thread.
-
-    Return:
-
-        enabled (bool)
-        course_id (str or None)
-    """
-    if is_forum_v2_disabled_globally():
-        return False, None
-
-    course_id = forum_api.get_course_id_by_comment(comment_id)
-    course_key = get_course_key(course_id)
-    return is_forum_v2_enabled(course_key), course_id
