+++ openedx/core/djangoapps/django_comment_common/comment_client/thread.py
@@ -2,20 +2,11 @@
 
 
 import logging
-import time
-import typing as t
 
 from eventtracking import tracker
 
-from django.core.exceptions import ObjectDoesNotExist
 from forum import api as forum_api
-from forum.api.threads import prepare_thread_api_response
-from forum.backend import get_backend
-from forum.backends.mongodb.threads import CommentThread
-from forum.utils import ForumV2RequestError
-from rest_framework.serializers import ValidationError
 
-from openedx.core.djangoapps.discussions.config.waffle import is_forum_v2_enabled, is_forum_v2_disabled_globally
 from . import models, settings, utils
 
 
@@ -169,7 +160,7 @@ class Thread(models.Model):
         request_params = utils.clean_forum_params(request_params)
         course_id = kwargs.get("course_id")
         if not course_id:
-            _, course_id = is_forum_v2_enabled_for_thread(self.id)
+            course_id = forum_api.get_course_id_by_thread(self.id)
         if user_id := request_params.get('user_id'):
             request_params['user_id'] = str(user_id)
         response = forum_api.get_thread(
@@ -225,125 +216,19 @@ class Thread(models.Model):
         )
         self._update_from_response(response)
 
-    @classmethod
-    def get_user_threads_count(cls, user_id, course_ids):
-        """
-        Returns threads and responses count of user in the given course_ids.
-        TODO: Add support for MySQL backend as well
-        """
-        query_params = {
-            "course_id": {"$in": course_ids},
-            "author_id": str(user_id),
-            "_type": "CommentThread"
-        }
-        return CommentThread()._collection.count_documents(query_params)  # pylint: disable=protected-access
 
-    @classmethod
-    def _delete_thread(cls, thread_id, course_id=None):
-        """
-        Deletes a thread
-        """
-        prefix = "<<Bulk Delete Thread>>"
-        backend = get_backend(course_id)()
-        try:
-            start_time = time.perf_counter()
-            thread = backend.validate_object("CommentThread", thread_id)
-            log.info(f"{prefix} Thread fetch {time.perf_counter() - start_time} sec")
-        except ObjectDoesNotExist as exc:
-            log.error("Forumv2RequestError for delete thread request.")
-            raise ForumV2RequestError(
-                f"Thread does not exist with Id: {thread_id}"
-            ) from exc
-
-        start_time = time.perf_counter()
-        backend.delete_comments_of_a_thread(thread_id)
-        log.info(f"{prefix} Delete comments of thread {time.perf_counter() - start_time} sec")
-
-        try:
-            start_time = time.perf_counter()
-            serialized_data = prepare_thread_api_response(thread, backend)
-            log.info(f"{prefix} Prepare response {time.perf_counter() - start_time} sec")
-        except ValidationError as error:
-            log.error(f"Validation error in get_thread: {error}")
-            raise ForumV2RequestError("Failed to prepare thread API response") from error
-
-        start_time = time.perf_counter()
-        backend.delete_subscriptions_of_a_thread(thread_id)
-        log.info(f"{prefix} Delete subscriptions {time.perf_counter() - start_time} sec")
-
-        start_time = time.perf_counter()
-        result = backend.delete_thread(thread_id)
-        log.info(f"{prefix} Delete thread {time.perf_counter() - start_time} sec")
-        if result and not (thread["anonymous"] or thread["anonymous_to_peers"]):
-            start_time = time.perf_counter()
-            backend.update_stats_for_course(
-                thread["author_id"], thread["course_id"], threads=-1
-            )
-            log.info(f"{prefix} Update stats {time.perf_counter() - start_time} sec")
-        return serialized_data
-
-    @classmethod
-    def delete_user_threads(cls, user_id, course_ids):
-        """
-        Deletes threads of user in the given course_ids.
-        TODO: Add support for MySQL backend as well
-        """
-        start_time = time.time()
-        query_params = {
-            "course_id": {"$in": course_ids},
-            "author_id": str(user_id),
-        }
-        threads_deleted = 0
-        threads = CommentThread().get_list(**query_params)
-        log.info(f"<<Bulk Delete>> Fetched threads for user {user_id} in {time.time() - start_time} seconds")
-        for thread in threads:
-            start_time = time.time()
-            thread_id = thread.get("_id")
-            course_id = thread.get("course_id")
-            if thread_id:
-                cls._delete_thread(thread_id, course_id=course_id)
-                threads_deleted += 1
-            log.info(f"<<Bulk Delete>> Deleted thread {thread_id} in {time.time() - start_time} seconds."
-                     f" Thread Found: {thread_id is not None}")
-        return threads_deleted
-
-
-def _url_for_flag_abuse_thread(thread_id):
-    return f"{settings.PREFIX}/threads/{thread_id}/abuse_flag"
-
-
-def _url_for_unflag_abuse_thread(thread_id):
-    return f"{settings.PREFIX}/threads/{thread_id}/abuse_unflag"
-
-
-def _url_for_pin_thread(thread_id):
-    return f"{settings.PREFIX}/threads/{thread_id}/pin"
-
-
-def _url_for_un_pin_thread(thread_id):
-    return f"{settings.PREFIX}/threads/{thread_id}/unpin"
-
-
-def is_forum_v2_enabled_for_thread(thread_id: str) -> tuple[bool, t.Optional[str]]:
-    """
-    Figure out whether we use forum v2 for a given thread.
-
-    This is a complex affair... First, we check the value of the DISABLE_FORUM_V2
-    setting, which overrides everything. If this setting does not exist, then we need to
-    find the course ID that corresponds to the thread ID. Then, we return the value of
-    the course waffle flag for this course ID.
-
-    Note that to fetch the course ID associated to a thread ID, we need to connect both
-    to mongodb and mysql. As a consequence, when forum v2 needs adequate connection
-    strings for both backends.
-
-    Return:
-
-        enabled (bool)
-        course_id (str or None)
-    """
-    if is_forum_v2_disabled_globally():
-        return False, None
-    course_id = forum_api.get_course_id_by_thread(thread_id)
-    course_key = utils.get_course_key(course_id)
-    return is_forum_v2_enabled(course_key), course_id
+def _clean_forum_params(params):
+    """Convert string booleans to actual booleans and remove None values from forum parameters."""
+    result = {}
+    for k, v in params.items():
+        if v is not None:
+            if isinstance(v, str):
+                if v.lower() == 'true':
+                    result[k] = True
+                elif v.lower() == 'false':
+                    result[k] = False
+                else:
+                    result[k] = v
+            else:
+                result[k] = v
+    return result
