+++ openedx/core/djangoapps/django_comment_common/comment_client/user.py
@@ -4,7 +4,6 @@
 from . import models, settings, utils
 from forum import api as forum_api
 from forum.utils import ForumV2RequestError, str_to_bool
-from openedx.core.djangoapps.discussions.config.waffle import is_forum_v2_enabled
 
 
 class User(models.Model):
@@ -38,17 +37,7 @@ class User(models.Model):
         """
         course_id = self.attributes.get("course_id")
         course_key = utils.get_course_key(course_id)
-        if is_forum_v2_enabled(course_key):
-            forum_api.mark_thread_as_read(self.id, source.id, course_id=str(course_id))
-        else:
-            params = {'source_type': source.type, 'source_id': source.id}
-            utils.perform_request(
-                'post',
-                _url_for_read(self.id),
-                params,
-                metric_action='user.read',
-                metric_tags=self._metric_tags + [f'target.type:{source.type}'],
-            )
+        forum_api.mark_thread_as_read(self.id, source.id, course_id=str(course_id))
 
     def follow(self, source, course_id=None):
         course_key = utils.get_course_key(self.attributes.get("course_id") or course_id)
@@ -110,31 +99,21 @@ class User(models.Model):
             query_params = {}
         if not self.course_id:
             raise utils.CommentClientRequestError("Must provide course_id when retrieving active threads for the user")
-        url = _url_for_user_active_threads(self.id)
         params = {'course_id': str(self.course_id)}
         params.update(query_params)
         course_key = utils.get_course_key(self.attributes.get("course_id"))
-        if is_forum_v2_enabled(course_key):
-            if user_id := params.get("user_id"):
-                params["user_id"] = str(user_id)
-            if page := params.get("page"):
-                params["page"] = int(page)
-            if per_page := params.get("per_page"):
-                params["per_page"] = int(per_page)
-            if count_flagged := params.get("count_flagged", False):
-                params["count_flagged"] = str_to_bool(count_flagged)
-            if not params.get("course_id"):
-                params["course_id"] = str(course_key)
-            response = forum_api.get_user_active_threads(**params)
-        else:
-            response = utils.perform_request(
-                'get',
-                url,
-                params,
-                metric_action='user.active_threads',
-                metric_tags=self._metric_tags,
-                paged_results=True,
-            )
+        if user_id := params.get("user_id"):
+            params["user_id"] = str(user_id)
+        if page := params.get("page"):
+            params["page"] = int(page)
+        if per_page := params.get("per_page"):
+            params["per_page"] = int(per_page)
+        if count_flagged := params.get("count_flagged", False):
+            params["count_flagged"] = str_to_bool(count_flagged)
+        if not params.get("course_id"):
+            params["course_id"] = str(course_key)
+        params = _clean_forum_params(params)
+        response = forum_api.get_user_active_threads(**params)
         return response.get('collection', []), response.get('page', 1), response.get('num_pages', 1)
 
     def subscribed_threads(self, query_params=None):
@@ -157,10 +136,8 @@ class User(models.Model):
             params["count_flagged"] = str_to_bool(count_flagged)
         if not params.get("course_id"):
             params["course_id"] = str(course_key)
-        if 'text' in params:
-            params.pop('text')
-        params = utils.clean_forum_params(params)
-        response = forum_api.get_user_subscriptions(**params)
+        params = _clean_forum_params(params)
+        response = forum_api.get_user_threads(**params)
         return utils.CommentClientPaginatedResult(
             collection=response.get('collection', []),
             page=response.get('page', 1),
@@ -169,7 +146,6 @@ class User(models.Model):
         )
 
     def _retrieve(self, *args, **kwargs):
-        url = self.url(action='get', params=self.attributes)
         retrieve_params = self.default_retrieve_params.copy()
         retrieve_params.update(kwargs)
 
@@ -183,116 +159,43 @@ class User(models.Model):
         if course_id:
             course_id = str(course_id)
             retrieve_params['course_id'] = course_id
-        course_key = utils.get_course_key(course_id) or utils.get_course_key(kwargs.get("course_key"))
 
-        if is_forum_v2_enabled(course_key):
-            group_ids = [retrieve_params['group_id']] if 'group_id' in retrieve_params else []
-            is_complete = retrieve_params['complete']
-            params = utils.clean_forum_params({
-                "user_id": self.attributes["id"],
-                "group_ids": group_ids,
-                "course_id": course_id,
-                "complete": is_complete
-            })
-            try:
-                response = forum_api.get_user(**params)
-            except ForumV2RequestError as e:
-                course_id = str(course_key)
-                self.save({"course_id": course_id})
-                response = forum_api.get_user(**params)
-        else:
-            try:
-                response = utils.perform_request(
-                    'get',
-                    url,
-                    retrieve_params,
-                    metric_action='model.retrieve',
-                    metric_tags=self._metric_tags,
-                )
-            except utils.CommentClientRequestError as e:
-                if e.status_code == 404:
-                    # attempt to gracefully recover from a previous failure
-                    # to sync this user to the comments service.
-                    self.save()
-                    response = utils.perform_request(
-                        'get',
-                        url,
-                        retrieve_params,
-                        metric_action='model.retrieve',
-                        metric_tags=self._metric_tags,
-                    )
-                else:
-                    raise
+        group_ids = [retrieve_params['group_id']] if 'group_id' in retrieve_params else None
+        is_complete = retrieve_params['complete']
+        params = _clean_forum_params({
+            "user_id": self.attributes["id"],
+            "group_ids": group_ids,
+            "course_id": course_id,
+            "complete": is_complete
+        })
+        try:
+            response = forum_api.get_user(**params)
+        except ForumV2RequestError as e:
+            self.save({"course_id": course_id})
+            response = forum_api.get_user(**params)
         self._update_from_response(response)
 
     def retire(self, retired_username):
         course_key = utils.get_course_key(self.attributes.get("course_id"))
-        if is_forum_v2_enabled(course_key):
-            forum_api.retire_user(user_id=self.id, retired_username=retired_username, course_id=str(course_key))
-        else:
-            url = _url_for_retire(self.id)
-            params = {'retired_username': retired_username}
-            utils.perform_request(
-                'post',
-                url,
-                params,
-                raw=True,
-                metric_action='user.retire',
-                metric_tags=self._metric_tags
-            )
+        forum_api.retire_user(user_id=self.id, retired_username=retired_username, course_id=str(course_key))
 
     def replace_username(self, new_username):
         course_key = utils.get_course_key(self.attributes.get("course_id"))
-        if is_forum_v2_enabled(course_key):
-            forum_api.update_username(user_id=self.id, new_username=new_username, course_id=str(course_key))
-        else:
-            url = _url_for_username_replacement(self.id)
-            params = {"new_username": new_username}
-
-            utils.perform_request(
-                'post',
-                url,
-                params,
-                raw=True,
-            )
+        forum_api.update_username(user_id=self.id, new_username=new_username, course_id=str(course_key))
 
 
-def _url_for_vote_comment(comment_id):
-    return f"{settings.PREFIX}/comments/{comment_id}/votes"
-
-
-def _url_for_vote_thread(thread_id):
-    return f"{settings.PREFIX}/threads/{thread_id}/votes"
-
-
-def _url_for_subscription(user_id):
-    return f"{settings.PREFIX}/users/{user_id}/subscriptions"
-
-
-def _url_for_user_active_threads(user_id):
-    return f"{settings.PREFIX}/users/{user_id}/active_threads"
-
-
-def _url_for_user_subscribed_threads(user_id):
-    return f"{settings.PREFIX}/users/{user_id}/subscribed_threads"
-
-
-def _url_for_read(user_id):
-    """
-    Returns cs_comments_service url endpoint to mark thread as read for given user_id
-    """
-    return f"{settings.PREFIX}/users/{user_id}/read"
-
-
-def _url_for_retire(user_id):
-    """
-    Returns cs_comments_service url endpoint to retire a user (remove all post content, etc.)
-    """
-    return f"{settings.PREFIX}/users/{user_id}/retire"
-
-
-def _url_for_username_replacement(user_id):
-    """
-    Returns cs_comments_servuce url endpoint to replace the username of a user
-    """
-    return f"{settings.PREFIX}/users/{user_id}/replace_username"
+def _clean_forum_params(params):
+    """Convert string booleans to actual booleans and remove None values from forum parameters."""
+    result = {}
+    for k, v in params.items():
+        if v is not None:
+            if isinstance(v, str):
+                if v.lower() == 'true':
+                    result[k] = True
+                elif v.lower() == 'false':
+                    result[k] = False
+                else:
+                    result[k] = v
+            else:
+                result[k] = v
+    return result
